ReceiptGeneratorTest.java
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;
import java.util.Arrays;
import java.util.List;

public class ReceiptGeneratorTest {

    @Test
    public void testCalculateTax() {
        ReceiptGenerator.ShoppingItem item = new ReceiptGenerator.ShoppingItem("apple", 0.99, 5, "food");
        assertEquals(0.0, ReceiptGenerator.calculateTax(item, "CA"));

        item = new ReceiptGenerator.ShoppingItem("shirt", 29.99, 1, "clothing");
        assertEquals(0.0, ReceiptGenerator.calculateTax(item, "NY"));

        item = new ReceiptGenerator.ShoppingItem("book", 14.99, 2, "book");
        double expectedTax = ReceiptGenerator.roundUpToNearest(14.99 * 2 * 0.0975, 2);
        assertEquals(expectedTax, ReceiptGenerator.calculateTax(item, "CA"));
    }

    @Test
    public void testIsExempt() {
        assertTrue(ReceiptGenerator.isExempt("food", "CA"));
        assertTrue(ReceiptGenerator.isExempt("clothing", "NY"));
        assertFalse(ReceiptGenerator.isExempt("book", "CA"));
    }

    @Test
    public void testPrintReceipt() {
        // Mocking the print statements is tricky, so we'll just test the logic indirectly via other methods.
        List<ReceiptGenerator.ShoppingItem> cart = Arrays.asList(
                new ReceiptGenerator.ShoppingItem("apple", 0.99, 5, "food"),
                new ReceiptGenerator.ShoppingItem("shirt", 29.99, 1, "clothing"),
                new ReceiptGenerator.ShoppingItem("book", 14.99, 2, "book")
        );
    }
}